requirements.txt
python-telegram-bot==22.0
dotenv==0.9.9
yt-dlp==2025.3.31
